Match icmpid in icmp_probe_match.
Remainder of Chris Johnson's patch from
http://seclists.org/nmap-dev/2013/q3/296.

Previously, Nmap could match replies that came in response to an
original probe, with one of the retransmissions of the probe. One effect
was that latency would be measured to be smaller than it should. Chris
Johnson summarized the problem and showed how to reproduce it at
http://seclists.org/nmap-dev/2013/q3/113.

git-svn-id: https://svn.nmap.org/nmap@31888 e0a8ed71-7df4-0310-8962-fdc924857419