Fix buffer overflow in parse_resolvconf()
String ipaddr was allocated without allowing space for the null
terminator, resulting in a 1-byte overflow. Caught with clang
-fsanitize=address

Also, fmt was being initialized with sizeof(ipaddr), which happened to
be correct, but should not necessarily be so. We don't care about the
size of the structure, but rather the length of an address in string
notation.

git-svn-id: https://svn.nmap.org/nmap@32524 e0a8ed71-7df4-0310-8962-fdc924857419