Fix NSEdoc generation problem
@field tag names in NSEdoc must be valid identifiers, so they cannot
contain "-". As a general rule, anything that needs to be quoted like
this: mytable["field-name"] is invalid. In this case, the ajp library
had a field called "status-line", which caused NSEdoc generation to fail
when it was finally documented. This change renames it to "status_line",
which should fix the issue.

git-svn-id: https://svn.nmap.org/nmap@32777 e0a8ed71-7df4-0310-8962-fdc924857419