Revert r32789 in favor of lib-level fixes
nmap.new_try() shouldn't be used in libraries. It results in Lua errors
being thrown that the script can't recover from without resorting to
pcall(). It has been replaced in proxy.lua with proper error handling
which did not require any changes to the scripts (http-open-proxy and
socks-open-proxy) that used it.

git-svn-id: https://svn.nmap.org/nmap@32791 e0a8ed71-7df4-0310-8962-fdc924857419