Fix man pages, corrupted from auto-generation
The docbook-to-nroff XSL pages turn indexterms into nroff comments, but
have a bug: If the closing indexterm element is not followed by
whitespace and cdata, then the following element or cdata is stuck on
the same line as the comment. Fixed this temporarily by introducing
whitespace between </indexterm> and following cdata or by moving the
indexterm elements after any other element (<command>, <term>, etc.)
that they reference.

git-svn-id: https://svn.nmap.org/nmap@33749 e0a8ed71-7df4-0310-8962-fdc924857419